backend/scripts/python/requirements.txt
deepface==0.0.93
opencv-python>=4.8.1.78
numpy==1.23.5
pandas>=2.1.4
Pillow>=9.5.0
requests>=2.31.0
tensorflow-cpu==2.12.0 